What Versions of Package Dependencies are Installed?

What Versions of Package Dependencies are Installed?

Usage

installedDependencies(package, recursive = TRUE)

Arguments

package

name of the package of which to check the dependencies

recursive

whether to look recursively for dependencies or only for the direct dependencies of package. Passed to packageDependencies, defaults to TRUE

Examples

installedDependencies(package = "kwb.package")
#>             Version  
#> Matrix      "1.4-1"  
#> R6          "2.5.1"  
#> askpass     "1.1"    
#> base64enc   "0.1-3"  
#> brew        "1.0-7"  
#> brio        "1.1.3"  
#> cachem      "1.0.6"  
#> callr       "3.7.0"  
#> cli         "3.3.0"  
#> clipr       "0.8.0"  
#> commonmark  "1.8.0"  
#> cpp11       "0.4.2"  
#> crayon      "1.5.1"  
#> credentials "1.3.2"  
#> curl        "4.3.2"  
#> desc        "1.4.1"  
#> devtools    "2.4.3"  
#> diffobj     "0.3.5"  
#> digest      "0.6.29" 
#> ellipsis    "0.3.2"  
#> evaluate    "0.15"   
#> fansi       "1.0.3"  
#> fastmap     "1.1.0"  
#> fs          "1.5.2"  
#> gert        "1.6.0"  
#> gh          "1.3.0"  
#> gitcreds    "0.1.1"  
#> glue        "1.6.2"  
#> grDevices   "4.2.0"  
#> graphics    "4.2.0"  
#> grid        "4.2.0"  
#> highr       "0.9"    
#> htmltools   "0.5.2"  
#> htmlwidgets "1.5.4"  
#> httr        "1.4.3"  
#> igraph      "1.3.1"  
#> ini         "0.3.1"  
#> jsonlite    "1.8.0"  
#> knitr       "1.39"   
#> kwb.utils   "0.13.0" 
#> lattice     "0.20-45"
#> lifecycle   "1.0.1"  
#> magrittr    "2.0.3"  
#> memoise     "2.0.1"  
#> methods     "4.2.0"  
#> mime        "0.12"   
#> mvbutils    "2.8.232"
#> networkD3   "0.4"    
#> openssl     "2.0.2"  
#> pillar      "1.7.0"  
#> pkgbuild    "1.3.1"  
#> pkgconfig   "2.0.3"  
#> pkgload     "1.2.4"  
#> praise      "1.0.0"  
#> prettyunits "1.1.1"  
#> processx    "3.6.0"  
#> ps          "1.7.0"  
#> purrr       "0.3.4"  
#> rappdirs    "0.3.3"  
#> rcmdcheck   "1.4.0"  
#> rematch2    "2.1.2"  
#> remotes     "2.4.2"  
#> rlang       "1.0.2"  
#> roxygen2    "7.2.0"  
#> rprojroot   "2.0.3"  
#> rstudioapi  "0.13"   
#> rversions   "2.1.1"  
#> sessioninfo "1.2.2"  
#> stats       "4.2.0"  
#> stringi     "1.7.6"  
#> stringr     "1.4.0"  
#> sys         "3.4"    
#> testthat    "3.1.4"  
#> tibble      "3.1.7"  
#> tools       "4.2.0"  
#> usethis     "2.1.6"  
#> utf8        "1.2.2"  
#> utils       "4.2.0"  
#> vctrs       "0.4.1"  
#> waldo       "0.4.0"  
#> whisker     "0.4"    
#> withr       "2.5.0"  
#> xfun        "0.31"   
#> xml2        "1.3.3"  
#> xopen       "1.0.0"  
#> yaml        "2.3.5"  
#> zip         "2.2.0"  
installedDependencies(package = "kwb.package", recursive = FALSE)
#>           Version  
#> devtools  "2.4.3"  
#> gh        "1.3.0"  
#> jsonlite  "1.8.0"  
#> kwb.utils "0.13.0" 
#> mvbutils  "2.8.232"
#> networkD3 "0.4"    
#> remotes   "2.4.2"  
#> withr     "2.5.0"